multiple clients added

+++ src/main/java/com/epam/training/designpatterns/fastfoodrestaurant/Chips.java
@@ -3,12 +3,12 @@ package com.epam.training.designpatterns.fastfoodrestaurant;
 public class Chips extends Product implements Consumable {
 
 	private final String NAME = "chips";
-	private int effect = 5;
+	private double effect = 5.0;
 
 	@Override
 	public void consumedBy(Client client) {
-		int clientHappiness = client.getHappiness();
-		client.setHappiness(clientHappiness * (100 + effect / 100));
+		double clientHappiness = client.getHappiness();
+		client.setHappiness(clientHappiness * ((100 + effect) / 100));
 	}
 
 	@Override
@@ -17,12 +17,12 @@ public class Chips extends Product implements Consumable {
 	}
 
 	@Override
-	public int getEffect() {
+	public double getEffect() {
 		return this.effect;
 	}
 
 	@Override
-	public void setEffect(int effect) {
+	public void setEffect(double effect) {
 		this.effect = effect;
 	}
